refactoring

+++ include/vertex.h
@@ -60,6 +60,6 @@ const std::vector<Vertex> quadVertices = {
   {{-0.5f, 1.0f, 0.0f}, {1.0f, 1.0f}}
 };
 
-const std::vector<uint16_t> quadIndices = {
+const std::vector<uint32_t> quadIndices = {
     0, 1, 2, 2, 3, 0
 };
